// packages/disk-moon-city/moonCity/types.ts
export enum BUILDING_KEY {
  ORE_STORAGE = 'ore_storage',
  REFINERY = 'refinery',
  POWER_STATION = 'power_station',
  ENERGY_TRANSMITTER = 'energy_transmitter',
  HOUSE = 'house',
  GREEN_HOUSE = 'green_house',
  FOOD_STORAGE = 'food_storage',
  VEHICLE_FACTORY = 'vehicle_factory',
  BARRACK = 'barrack',
  WEAPON_FACTORY = 'weapon_factory',
  VAULT = 'vault',
  SHIELD_GENERATOR = 'shield_generator'
}
// ...
export enum RESOURCE_TYPE {
  NONE = 'none',
  MINERAL_ORE = 'mineral_ore',
  ENERGY_CELL = 'energy_cell',
  FOOD = 'food',
  ENERGY = 'energy',
  CREDITS = 'credits',
  SHIELD_ENERGY = 'shield_energy'
}

export enum STORAGE_TYPE {
  NONE = 'none',
  CREDITS = 'credits',
  MINERAL_ORE = 'mineral_ore',
  ENERGY_CELL = 'energy_cell',
  FOOD = 'food',
  ENERGY = 'energy',
  HUMAN = 'human',
  EMPLOYEE = 'employee',
  SECURITY_SERVICE = 'security_service',
  SOLDIER = 'soldiers',
  MERCENARY = 'mercenary',
  ENERGY_TRANSFER = 'energy_transfer',
  SHIELD_ENERGY = 'shield_energy'
}
// ...
export enum LINE_GROUP {
  VEHICLE = 'vehicle',
  GENERAL = 'general',
  COST = 'cost',
  INCOME = 'income',
  ATTACK = 'attack'
}